+++ pkg/googlesheets/googlesheets.go
@@ -1,18 +1,21 @@
 package googlesheets
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"strings"
 	"time"
 
-	"context"
+	"github.com/grafana/google-sheets-datasource/pkg/models"
 
 	"github.com/araddon/dateparse"
 	"github.com/davecgh/go-spew/spew"
-	"github.com/grafana/google-sheets-datasource/pkg/models"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/patrickmn/go-cache"
+	"google.golang.org/api/googleapi"
 	"google.golang.org/api/sheets/v4"
 )
 
@@ -22,7 +25,7 @@ type GoogleSheets struct {
 }
 
 // Query queries a spreadsheet and returns a corresponding data frame.
-func (gs *GoogleSheets) Query(ctx context.Context, refID string, qm *models.QueryModel, config *models.DatasourceSettings, timeRange backend.TimeRange) (dr backend.DataResponse) {
+func (gs *GoogleSheets) Query(ctx context.Context, refID string, qm *models.QueryModel, config models.DatasourceSettings, timeRange backend.TimeRange) (dr backend.DataResponse) {
 	client, err := NewGoogleClient(ctx, config)
 	if err != nil {
 		dr.Error = fmt.Errorf("unable to create Google API client: %w", err)
@@ -30,13 +33,13 @@ func (gs *GoogleSheets) Query(ctx context.Context, refID string, qm *models.Quer
 	}
 
 	// This result may be cached
-	data, meta, err := gs.getSheetData(client, qm)
+	sheetData, meta, err := gs.getSheetData(client, qm)
 	if err != nil {
 		dr.Error = err
 		return
 	}
 
-	frame, err := gs.transformSheetToDataFrame(data, meta, refID, qm)
+	frame, err := gs.transformSheetToDataFrame(sheetData, meta, refID, qm)
 	if err != nil {
 		dr.Error = err
 		return
@@ -47,7 +50,7 @@ func (gs *GoogleSheets) Query(ctx context.Context, refID string, qm *models.Quer
 	if qm.UseTimeFilter {
 		timeIndex := findTimeField(frame)
 		if timeIndex >= 0 {
-			frame, dr.Error = frame.FilterRowsByField(timeIndex, func(i interface{}) (bool, error) {
+			frame, dr.Error = frame.FilterRowsByField(timeIndex, func(i any) (bool, error) {
 				val, ok := i.(*time.Time)
 				if !ok {
 					return false, fmt.Errorf("invalid time column: %s", spew.Sdump(i))
@@ -64,7 +67,7 @@ func (gs *GoogleSheets) Query(ctx context.Context, refID string, qm *models.Quer
 }
 
 // GetSpreadsheets gets spreadsheets from the Google API.
-func (gs *GoogleSheets) GetSpreadsheets(ctx context.Context, config *models.DatasourceSettings) (map[string]string, error) {
+func (gs *GoogleSheets) GetSpreadsheets(ctx context.Context, config models.DatasourceSettings) (map[string]string, error) {
 	client, err := NewGoogleClient(ctx, config)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create Google API client: %w", err)
@@ -84,10 +87,10 @@ func (gs *GoogleSheets) GetSpreadsheets(ctx context.Context, config *models.Data
 }
 
 // getSheetData gets grid data corresponding to a spreadsheet.
-func (gs *GoogleSheets) getSheetData(client client, qm *models.QueryModel) (*sheets.GridData, map[string]interface{}, error) {
+func (gs *GoogleSheets) getSheetData(client client, qm *models.QueryModel) (*sheets.GridData, map[string]any, error) {
 	cacheKey := qm.Spreadsheet + qm.Range
 	if item, expires, found := gs.Cache.GetWithExpiration(cacheKey); found && qm.CacheDurationSeconds > 0 {
-		return item.(*sheets.GridData), map[string]interface{}{
+		return item.(*sheets.GridData), map[string]any{
 			"hit":     true,
 			"expires": expires.Unix(),
 		}, nil
@@ -95,27 +98,38 @@ func (gs *GoogleSheets) getSheetData(client client, qm *models.QueryModel) (*she
 
 	result, err := client.GetSpreadsheet(qm.Spreadsheet, qm.Range, true)
 	if err != nil {
-		return nil, nil, err
+		if apiErr, ok := err.(*googleapi.Error); ok {
+			// Handle API-specific errors
+			if apiErr.Code == 404 {
+				return nil, nil, errors.New("spreadsheet not found")
+			}
+			if apiErr.Message != "" {
+				log.DefaultLogger.Error("Google API Error: " + apiErr.Message)
+				return nil, nil, fmt.Errorf("Google API Error %d", apiErr.Code)
+			}
+			log.DefaultLogger.Error(apiErr.Error())
+			return nil, nil, errors.New("unknown API error")
+		}
 	}
 
 	if result.Properties.TimeZone != "" {
 		loc, err := time.LoadLocation(result.Properties.TimeZone)
 		if err != nil {
-			backend.Logger.Warn("could not load timezone from spreadsheet: %w", err)
+			log.DefaultLogger.Warn("could not load timezone from spreadsheet: %w", err)
 		} else {
 			time.Local = loc
 		}
 	}
 
-	data := result.Sheets[0].Data[0]
+	sheetData := result.Sheets[0].Data[0]
 	if qm.CacheDurationSeconds > 0 {
-		gs.Cache.Set(cacheKey, data, time.Duration(qm.CacheDurationSeconds)*time.Second)
+		gs.Cache.Set(cacheKey, sheetData, time.Duration(qm.CacheDurationSeconds)*time.Second)
 	}
 
-	return data, map[string]interface{}{"hit": false}, nil
+	return sheetData, map[string]any{"hit": false}, nil
 }
 
-func (gs *GoogleSheets) transformSheetToDataFrame(sheet *sheets.GridData, meta map[string]interface{}, refID string, qm *models.QueryModel) (*data.Frame, error) {
+func (gs *GoogleSheets) transformSheetToDataFrame(sheet *sheets.GridData, meta map[string]any, refID string, qm *models.QueryModel) (*data.Frame, error) {
 	columns, start := getColumnDefinitions(sheet.RowData)
 	warnings := []string{}
 
@@ -146,13 +160,13 @@ func (gs *GoogleSheets) transformSheetToDataFrame(sheet *sheets.GridData, meta m
 		if column.HasMixedTypes() {
 			warning := fmt.Sprintf("Multiple data types found in column %q. Using string data type", column.Header)
 			warnings = append(warnings, warning)
-			backend.Logger.Warn(warning)
+			log.DefaultLogger.Warn(warning)
 		}
 
 		if column.HasMixedUnits() {
 			warning := fmt.Sprintf("Multiple units found in column %q. Formatted value will be used", column.Header)
 			warnings = append(warnings, warning)
-			backend.Logger.Warn(warning)
+			log.DefaultLogger.Warn(warning)
 		}
 	}
 
@@ -178,14 +192,14 @@ func (gs *GoogleSheets) transformSheetToDataFrame(sheet *sheets.GridData, meta m
 	meta["spreadsheetId"] = qm.Spreadsheet
 	meta["range"] = qm.Range
 	frame.Meta = &data.FrameMeta{Custom: meta}
-	backend.Logger.Debug("frame.Meta: %s", spew.Sdump(frame.Meta))
+	log.DefaultLogger.Debug("frame.Meta: %s", spew.Sdump(frame.Meta))
 	return frame, nil
 }
 
 // timeConverter handles sheets TIME column types.
 var timeConverter = data.FieldConverter{
 	OutputFieldType: data.FieldTypeNullableTime,
-	Converter: func(i interface{}) (interface{}, error) {
+	Converter: func(i any) (any, error) {
 		var t *time.Time
 		cellData, ok := i.(*sheets.CellData)
 		if !ok {
@@ -202,7 +216,7 @@ var timeConverter = data.FieldConverter{
 // stringConverter handles sheets STRING column types.
 var stringConverter = data.FieldConverter{
 	OutputFieldType: data.FieldTypeNullableString,
-	Converter: func(i interface{}) (interface{}, error) {
+	Converter: func(i any) (any, error) {
 		var s *string
 		cellData, ok := i.(*sheets.CellData)
 		if !ok {
@@ -215,7 +229,7 @@ var stringConverter = data.FieldConverter{
 // numberConverter handles sheets STRING column types.
 var numberConverter = data.FieldConverter{
 	OutputFieldType: data.FieldTypeNullableFloat64,
-	Converter: func(i interface{}) (interface{}, error) {
+	Converter: func(i any) (any, error) {
 		cellData, ok := i.(*sheets.CellData)
 		if !ok {
 			return nil, fmt.Errorf("expected type *sheets.CellData, but got %T", i)