+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export { QueryEditor } from './QueryEditor';
 export { ConfigEditor } from './ConfigEditor';
-export { JWTConfig } from './JWTConfig';
 export { MetaInspector } from './MetaInspector';